WIP for moving sort functionality into custom hook
