+++ packages/react-storage/src/components/StorageBrowser/index.ts
@@ -1,1 +1,2 @@
 export { default as createStorageBrowser } from './createStorageBrowser';
+export { useSortHandler } from './hooks/useSortHandler';
