+++ packages/react-storage/src/components/index.ts
@@ -1,4 +1,4 @@
-export { createStorageBrowser } from './StorageBrowser';
+export { createStorageBrowser, useSortHandler } from './StorageBrowser';
 export { StorageImage, StorageImageProps } from './StorageImage';
 export {
   StorageManager,
