+++ packages/react-storage/src/index.ts
@@ -10,4 +10,5 @@ export {
   FilePickerProps,
   FileListHeaderProps,
   FileListFooterProps,
+  useSortHandler,
 } from './components';
